requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
wandb>=0.15.0
accelerate>=0.20.0
evaluate>=0.4.0
nltk>=3.8.1
spacy>=3.6.0
textblob>=0.17.1
demoji>=1.1.0
emot>=3.1
vaderSentiment>=3.3.2
wordcloud>=1.9.0
plotly>=5.15.0
emoji>=2.2.0
contractions>=0.1.73
autocorrect>=2.6.1
gensim>=4.3.0
pyyaml>=6.0
tensorboard>=2.13.0
jupyter>=1.0.0
ipywidgets>=8.0.0
gradio>=3.35.0
